feat: Introduce commodity management and refine currency handling

This pull request includes changes for two main enhancements:

1. Support for Commodity and Specification Management:
- Enable the creation of products along with package options, and manage
their lifecycle.
- Integrate with Compute Nest services to bind products with their
respective services.
- Introduce functionality to activate and deactivate products for public
listing.

2. Refactor Currency Representation:
- Transition all monetary values in the codebase to be represented using
BigDecimal, ensuring precision.
- Modify the underlying storage mechanism for currency-related fields,
moving from double to long type.
     This allows for storage of monetary amounts in minor units (cents).
- Update front-end display logic to convert long-type monetary values
from cents to yuan for user presentation.

These combined changes create a cohesive feature set for managing
commodities and offer accurate currency handling across the platform.

+++ boost.common/src/main/java/org/example/common/adapter/BaseAlipayClient.java
@@ -18,6 +18,8 @@ package org.example.common.adapter;
 import com.alipay.api.response.AlipayTradeQueryResponse;
 import org.example.common.config.AlipayConfig;
 
+import java.math.BigDecimal;
+
 public interface BaseAlipayClient {
 
     /**
@@ -42,7 +44,7 @@ public interface BaseAlipayClient {
      * @param outTradeNo out trade number
      * @return String
      */
-    String createTransaction(Double totalAmount, String subject, String outTradeNo);
+    String createTransaction(BigDecimal totalAmount, String subject, String outTradeNo);
 
     /**
      * Order Refund.
@@ -51,7 +53,7 @@ public interface BaseAlipayClient {
      * @param refundRequestId Refund Request Id
      * @return {@link Boolean}
      */
-    Boolean refundOrder(String orderId, Double refundAmount, String refundRequestId);
+    Boolean refundOrder(String orderId, BigDecimal refundAmount, String refundRequestId);
 
     /**
      * Close order.
